requirements.txt
Flask
google-generativeai
Pillow
python-dotenv
pymongo[srv]
gunicorn
werkzeug